Move RunJavaScript activity to Elsa.JavaScript
