+++ src/modules/Elsa.JavaScript/Elsa.JavaScript.csproj
@@ -13,12 +13,16 @@
 
     <ItemGroup>
         <ProjectReference Include="..\Elsa.Expressions\Elsa.Expressions.csproj" />
-        <ProjectReference Include="..\Elsa.Workflows.Core\Elsa.Workflows.Core.csproj" />
+        <ProjectReference Include="..\Elsa.Workflows.Management\Elsa.Workflows.Management.csproj" />
     </ItemGroup>
 
     <ItemGroup>
         <PackageReference Include="System.Linq.Async" Version="6.0.1" />
-        <PackageReference Include="Jint" Version="3.0.0-beta-2052" />
+        <PackageReference Include="Jint" Version="3.0.0-beta-2053" />
+    </ItemGroup>
+
+    <ItemGroup>
+      <Folder Include="Activities\" />
     </ItemGroup>
 
 </Project>
